% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewpoint_to_humidity.r
\name{dewpoint_to_humidity}
\alias{dewpoint_to_humidity}
\title{Relative Humidity}
\usage{
dewpoint_to_humidity(dp = NA, t = NA)
}
\arguments{
\item{dp}{= Dewpoint    (degrees celsius)}

\item{t}{= Temperature mean    (degrees celcius)}
}
\value{
A vector containing the daily vp
}
\description{
This function computes RH from dewpoint
}
