% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_priestley_taylor.r
\name{pet_priestley_taylor}
\alias{pet_priestley_taylor}
\title{Potential Evapotranspiration}
\usage{
pet_priestley_taylor(indat)
}
\arguments{
\item{indat}{= dataframe containing the input variables}
}
\value{
A vector containing the daily Potential EvapoTranspiration (mm/day)
}
\description{
This function takes a dataframe object as input and computes PET using Makkink's method.
}
\examples{
pet_priestley_taylor(indat)

Where the dataframe "indat" contains the following variables:

tmin  = Temperature min    (degrees Celsius)
tmax  = Temperature max    (degrees Celsius)
only one of the following variables (vp OR rh OR dp) are to be passed
vp  = Vapour pressure    (hPa)
rh =  Relative humidity  (\%)
dp =  Dewpoint           (degrees Celsius)
rs  = radiation          (MJ/m^2)
Also rquired is the latitude and longitude (for land sea mask)
lat = Latitude  (degrees)
lons = Longitude (degrees)

Missing values should be converted to NA

}
