% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_penmon.r
\name{vapour_pressure}
\alias{vapour_pressure}
\title{Vapour pressure}
\usage{
vapour_pressure(tm, vp)
}
\arguments{
\item{tm}{= Temperature    (degrees Celcious)}

\item{vp}{= Vapour pressure    (hPa)}
}
\value{
A vector containing the daily vp
}
\description{
This function computes vapour pressure from temperature
}
